Move FormattingOptions to its own file and remove duplication.
